requirements.txt
Django
djangorestframework
Pillow
psycopg
qrcode
reportlab
shortuuid
django-simple-history
requests
aiohttp
gunicorn
pandas
openpyxl
tenacity
cairosvg
schedule
langchain
langchain-google-genai 
pillow
langchain-experimental
